// src/payload-types.ts
export interface IeeeProject {
  id: number;
  name: string;
  description: string;
  logo: number | Media;
  websiteUrl?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * Max media upload size: 1MB. Please optimize your images before uploading.
 *
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  /**
   * If left empty, this will be auto-populated from the filename.
   */
  alt?: string | null;
  /**
   * Categorizing media helps in organizing the library and Cloudinary folders.
   */
  category?:
    | (
        | 'ypsl'
        | 'executive-committees'
        | 'standing-committees'
        | 'ai-driven-sri-lanka'
        | 'sl-inspire'
        | 'lets-talk'
        | 'insl'
        | 'studpro'
        | 'y2npro'
        | 'merch-payslips'
        | 'others'
        | 'merchants'
        | 'ieee-projects'
        | 'events'
      )
    | null;
  cloudinaryUrl?: string | null;
  cloudinaryPublicId?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}